+++ Addons/Uniforms/hpp/CfgVehicles.hpp
@@ -1,0 +1,85 @@
+class CfgVehicles
+{
+    class OPTRE_UNSC_Army_Soldier_BLK_SlimLeg;
+    class ONI_Recon_Army_Soldier_Blk_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_black_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_black_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_Blk_Slim";
+    };    
+    class ONI_Recon_Army_Soldier_BlkUrb_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_black_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_urban_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_BlkUrb_Slim";
+    };
+    class ONI_Recon_Army_Soldier_BlkMar_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_black_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_marine_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_BlkMar_Slim";
+    }; 
+    class ONI_Recon_Army_Soldier_BlkWdl_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_black_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_woodland_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_BlkWdl_Slim";
+    };
+    class ONI_Recon_Army_Soldier_UrbBlk_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_urban_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_black_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_UrbBlk_Slim";
+    };   
+    class ONI_Recon_Army_Soldier_Urb_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_urban_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_urban_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_Urb_Slim";
+    };  
+    //Why am i doing this?
+    class ONI_Recon_Army_Soldier_UrbMar_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_urban_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_marine_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_UrbMar_Slim";
+    };
+    class ONI_Recon_Army_Soldier_UrbWdl_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_urban_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_woodland_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_UrbWdl_Slim";
+    }; 
+    class ONI_Recon_Army_Soldier_MarBlk_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_marine_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_black_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_MarBlk_Slim";
+    };       
+    class ONI_Recon_Army_Soldier_MarUrb_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_marine_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_urban_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_MarUrb_Slim";
+    }; 
+    class ONI_Recon_Army_Soldier_Mar_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_marine_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_marine_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_Mar_Slim";
+    }; 
+    class ONI_Recon_Army_Soldier_MarWdl_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_marine_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_woodland_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_MarWdl_Slim";
+    };
+    class ONI_Recon_Army_Soldier_WdlBlk_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_woodland_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_black_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_WdlBlk_Slim";
+    };   
+    class ONI_Recon_Army_Soldier_WdlUrb_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_woodland_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_urban_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_WdlUrb_Slim";
+    };  
+    class ONI_Recon_Army_Soldier_WdlMar_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_woodland_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_marine_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_WdlMar_Slim";
+    };  
+    class ONI_Recon_Army_Soldier_Wdl_Slim: OPTRE_UNSC_Army_Soldier_BLK_SlimLeg
+    {
+        hiddenSelectionsTextures[] = {"\x\VS_ONI\Uniforms\data\ONI_uniform_a_woodland_co.paa","\x\VS_ONI\Uniforms\data\ONI_uniform_b_woodland_co.paa"};
+        uniformClass = "ONI_Recon_Uniform_Wdl_Slim";
+    };     
+};
