+++ Addons/Uniforms/hpp/CfgWeapons.hpp
@@ -1,0 +1,134 @@
+class CfgWeapons
+{
+    class ItemInfo;
+    class OPTRE_UNSC_Army_Uniform_BLK_SlimLeg;
+    class ONI_Recon_Uniform_Blk_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Black)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_Blk_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_BlkUrb_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Black/Urban)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_BlkUrb_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_BlkMar_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Black/Marine)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_BlkMar_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_BlkWdl_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Black/Woodland)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_BlkWdl_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_UrbBlk_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Urban/Black)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_UrbBlk_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_Urb_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Urban)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_Urb_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_UrbMar_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Urban/Marine)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_UrbMar_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_UrbWdl_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Urban/Woodland)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_UrbWdl_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_MarBlk_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Marine/Black)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_MarBlk_Slim";
+        };
+    };
+//This is just a lot of boring Ctrl+C -> Ctrl+V and making a few small changes, PS Drink water 
+    class ONI_Recon_Uniform_MarUrb_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Marine/Urban)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_MarUrb_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_Mar_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Marine)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_Mar_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_MarWdl_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Marine/Woodland)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_MarWdl_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_WdlBlk_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Woodland/Black)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_WdlBlk_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_WdlUrb_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Woodland/Urban)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_WdlUrb_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_WdlMar_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Woodland/Marine)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_WdlMar_Slim";
+        };
+    };
+    class ONI_Recon_Uniform_Wdl_Slim : OPTRE_UNSC_Army_Uniform_BLK_SlimLeg
+    {
+        displayName = "[ONI] BDU (Woodland)";
+        class ItemInfo: ItemInfo
+        {
+            uniformClass = "ONI_Recon_Army_Soldier_Wdl_Slim";
+        };
+    };
+};
