+++ Addons/Uniforms/hpp/XtdGearInfos.hpp
@@ -1,0 +1,84 @@
+class XtdGearInfos {
+	class CfgWeapons {
+		class ONI_Recon_Uniform_Blk_Slim {
+            model = "ONI_BDU";
+            Shirt = "A";
+            Trousers = "A";
+        };
+		class ONI_Recon_Uniform_BlkUrb_Slim {
+            model = "ONI_BDU";
+            Shirt = "A";
+            Trousers = "B";
+        };
+		class ONI_Recon_Uniform_BlkMar_Slim {
+            model = "ONI_BDU";
+            Shirt = "A";
+            Trousers = "C";
+        };
+		class ONI_Recon_Uniform_BlkWdl_Slim {
+            model = "ONI_BDU";
+            Shirt = "A";
+            Trousers = "D";
+        };
+		class ONI_Recon_Uniform_UrbBlk_Slim {
+            model = "ONI_BDU";
+            Shirt = "B";
+            Trousers = "A";
+        };
+		class ONI_Recon_Uniform_Urb_Slim {
+            model = "ONI_BDU";
+            Shirt = "B";
+            Trousers = "B";
+        };
+		class ONI_Recon_Uniform_UrbMar_Slim {
+            model = "ONI_BDU";
+            Shirt = "B";
+            Trousers = "C";
+        };
+		class ONI_Recon_Uniform_UrbWdl_Slim {
+            model = "ONI_BDU";
+            Shirt = "B";
+            Trousers = "D";
+        };
+		class ONI_Recon_Uniform_MarBlk_Slim {
+            model = "ONI_BDU";
+            Shirt = "C";
+            Trousers = "A";
+        };
+		class ONI_Recon_Uniform_MarUrb_Slim {
+            model = "ONI_BDU";
+            Shirt = "C";
+            Trousers = "B";
+        };
+		class ONI_Recon_Uniform_Mar_Slim {
+            model = "ONI_BDU";
+            Shirt = "C";
+            Trousers = "C";
+        };
+		class ONI_Recon_Uniform_MarWdl_Slim {
+            model = "ONI_BDU";
+            Shirt = "C";
+            Trousers = "D";
+        };
+		class ONI_Recon_Uniform_WdlBlk_Slim {
+            model = "ONI_BDU";
+            Shirt = "D";
+            Trousers = "A";
+        };
+		class ONI_Recon_Uniform_WdlUrb_Slim {
+            model = "ONI_BDU";
+            Shirt = "D";
+            Trousers = "B";
+        };
+		class ONI_Recon_Uniform_WdlMar_Slim {
+            model = "ONI_BDU";
+            Shirt = "D";
+            Trousers = "C";
+        };
+		class ONI_Recon_Uniform_Wdl_Slim {
+            model = "ONI_BDU";
+            Shirt = "D";
+            Trousers = "D";
+        };
+	};
+};
