+++ Addons/Uniforms/hpp/XtdGearModels.hpp
@@ -1,0 +1,27 @@
+class XtdGearModels {
+    class CfgWeapons {
+        class ONI_BDU {
+            label = "[ONI] BDU";
+            author = "Viking Studios";
+            options[] = {"Shirt","Trousers"};
+            class Shirt {
+                changeingame = 0;
+                values[] = {"A","B","C","D"};
+
+                class A { label = "Black"; };
+                class B { label = "Urban"; };
+                class C { label = "Marine"; };
+                class D { label = "Woodland"; };
+            };
+            class Trousers {
+                changeingame = 0;
+                values[] = {"A","B","C","D"};
+
+                class A { label = "Black"; };
+                class B { label = "Urban"; };
+                class C { label = "Marine"; };
+                class D { label = "Woodland"; };
+            };
+        };
+    };
+};
